feature - clipRasterByPoly
last commit of clipRasterByPoly before replacing with contents of clipRasterByPoly2, a temporary clean slate that I finished. If old behavior or notes are needed, check prior commit of clipRasterByPoly before this major rewrite.

Otherwise, first commit of utility functions to support clipRasterByPoly and overall clean up of libraster and libspatial.