fix(toolbox): Fix case where toolboxes.library is empty.
On other mac, no libraries are set, so the toolboxes.libary column did not exist. Tried to add a column of empty "" and cellstr but writetable converts those to nan. For now, readtable converts to string which converts that column to missing. Not sure what will happen when I add an entry with a library.