refactor(parsegraphics): Allow graphics in any position, see notes.
This is a major refactor to allow input graphics objects to exist in any position in varargin rather than the first position. In doing so, nearly the entire function was rewritten. The "parent, object" check was moved first because otherwise, if both a lone object and a parent object pair are supplied, both are removed by the check for a lone object, which previously came first. Now the parent object pair is removed followed by a check for a lone object. Note functions isgraphicslike and isimage were added to liblogic in the previous commit. The former encapsulates the check for a parent, object pair, which is broader than the isaxis and isfig checks and includes numeric double scalar handles.