fix(isfig/isaxis): Separate pre/post r2014b checks.
In pre-r2014b it may not be possible to distinguish a numeric scalar from an open figure with the same figure number. The pre/post if-else means this case only arises if the function is called in a pre-2014b environment. Also removed use of "isgraphics" in pre-2014b case.