fix(listfiles): Cast folderlist to column, use arraymap, and fix cell concat.
Since folderlist is a string array, use arraymap. Cast to column so list is a column when folderlist is not scalar, and in this case, list will be non-scalar cell array with one element per subfolder containing files in each subfolder, so add the "aslist" logical check to concat them into one long list.