fix(annualdMdt): Ensure correctness for left/right/none padding.
The long awaited fix to this function ensures bimonthly averages are computed correctly regardless of whether the left, right, both, or neither side is padded, puts the bimonthly averages into the second of the two months, and puts the annual differences into the start month/year of the annual period. This aligns with the convention used in waterBalanceAnalysis and the "retime" function with option "yearly" ie annual periods are registered to Jan1 of the year. While registering annual differences to Dec of the year makes indexing consistent between calendar and water years, overall its confusing and error prone.