feat(parsePolygonNames)!: Redesign for parsePolygons; add expectedNames, see notes.
Redesigned and simplified the inputs so the first one is either a struct/table with a name field or the list of names, this way the function can be called directly from parsePolygons for the case where P is a shapefile, or with a list of names from wherever, and parsedPolygons is an optional input. Also made it analogous to parseFieldNames, so basically the difference is in parsing the "name" field rather than the fieldnames themselves.