feat(parsePolygons)!: Change output order; parse and return polygonNames.
Call the refactored parsePolygonNames to enable moving "inputP" and "wasfileP" outputs to the end and "polygonNames" up to the fourth output which will be more useful in general. This also requires passing polygonNames as an input and only parsing polygonNames if its empty, which provides more flexibility when calling this function multiple times within another function where the names are already parsed.