fix(mkproject): Set toolbox-related success flags true
These flags are false by default indicating a failure to successfully copy the toolbox template and replace the +tbx prefix. But if maketoolbox is false, it is necessary to set these flags to true before the success checks are made later, otherwise the function would issue a warning and exit.